% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGLM.R
\name{coef.spglm}
\alias{coef.spglm}
\alias{vcov.spglm}
\alias{logLik.spglm}
\title{Extracting components of fitted \code{spglm} object}
\usage{
\method{coef}{spglm}(object, beta = TRUE, f0 = FALSE, ...)

\method{vcov}{spglm}(object, beta = TRUE, f0 = FALSE, ...)

\method{logLik}{spglm}(object, ...)
}
\arguments{
\item{object}{fitted model of class \code{spglm}}

\item{beta}{logical, whether the regression parameter estimates / variances should be returned}

\item{f0}{logical, whether the backbone density estimates / variances should be returned}

\item{...}{not used, present fo consistency with generics}
}
\description{
Extracting components of fitted \code{spglm} object
}
