% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgldrm.R
\name{gldrm.control}
\alias{gldrm.control}
\title{Main optimization function}
\usage{
gldrm.control(
  eps = 1e-10,
  maxiter = 100,
  returnfTiltMatrix = TRUE,
  returnf0ScoreInfo = FALSE,
  print = FALSE,
  betaStart = NULL,
  f0Start = NULL
)
}
\description{
This function is called by the main \code{gldrm} function.
}
\keyword{internal}
