% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGLM.R
\name{predict.spglm}
\alias{predict.spglm}
\title{Predict methods for SPGLM fits}
\usage{
\method{predict}{spglm}(
  object,
  newdata = NULL,
  type = c("mean", "prob", "tilt", "lp"),
  newn = NULL,
  newevents = NULL,
  ...
)
}
\arguments{
\item{object}{fitted model of class \code{spglm}}

\item{newdata}{optionally, a data frame in which to look for covariate values for prediction.
If NULL, the original data set is used}

\item{type}{the type of prediction requested. The default is "mean", the mean event probability;
"prob" requests the probability of the observation (given number of responses with given cluster size),
"tilt" returns the tilting parameter which achieves the modeled mean from the baseline distribution;
and "lp" requests the linear predictor.}

\item{newn}{if \code{newdata} is provided and \code{type="prob"}, an integer or integer vector specifying the clustersize for the predictors}

\item{newevents}{if \code{newdata} is provided and \code{type="prob"}, an integer or integer vector specifying the
number of events for the predictions}

\item{...}{not used, present for consistency with generic}
}
\description{
Obtains predictions from a fitted semi-parametric generalized linear model. Note that \code{offset}
and \code{weight} terms are not implemented for predicting from new data.
}
