% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGLM.R
\name{ran.spglm}
\alias{ran.spglm}
\title{Simulate data that follows an SPGLM}
\usage{
ran.spglm(n, means, q0)
}
\arguments{
\item{n}{vector of cluster-sizes for the simulated data.}

\item{means}{vector of mean event probability for each cluster. Values should be between 0 and 1.}

\item{q0}{the backbone probability distribution of number of events in a cluster of maximal size.}
}
\value{
a data frame with columns \code{Mean}, \code{ClusterSize}, \code{NResp}
}
\description{
Simulate data that follows an SPGLM
}
