% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGLM.R
\name{spglm}
\alias{spglm}
\title{Fit semi-parametric GLM}
\usage{
spglm(
  formula,
  data,
  subset,
  weights,
  link = "logit",
  start = NULL,
  control = list(eps = 0.001, maxit = 100),
  ...
)
}
\arguments{
\item{formula}{a one-sided formula of the form \code{cbind(r, s) ~ predictors} where \code{r} and \code{s} give the number of responses and non-responses within each cluster, respectively (so the cluster size is \code{r+s}), and \code{predictors} describes the covariates.}

\item{data}{an optional matrix or data frame containing the variables in the formula \code{formula}. By default the variables are taken from \code{environment(formula).}}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{link}{a link function for the mean.}

\item{start}{an optional list with elements named \code{beta}, \code{q}, and/or \code{mu1} giving starting values for estimation. If none or only some are specified, starting values are selected automatically.}

\item{control}{a list with parameters controlling the algorithm.}

\item{weight}{an optional vector specifying observation weights.}
}
\value{
an object of class \code{spglm} with the fitted model.
}
\description{
Fit semi-parametric GLM
}
