% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPreg.R
\name{sprr}
\alias{sprr}
\title{Fit semi-parametric relative risk model}
\usage{
sprr(
  formula,
  data,
  subset,
  weights,
  link = "cloglog",
  mu1 = NULL,
  start = NULL,
  control = list(eps = 0.001, maxit = 100),
  ...
)
}
\arguments{
\item{formula}{a one-sided formula of the form \code{cbind(r, s) ~ predictors} where \code{r} and \code{s} give the number of responses and non-responses within each cluster, respectively (so the cluster size is \code{r+s}), and \code{predictors} describes the covariates.}

\item{data}{an optional matrix or data frame containing the variables in the formula \code{formula}. By default the variables are taken from \code{environment(formula).}}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{link}{a link function for the binomial distribution, the inverse of which models the covariate effects.}

\item{mu1}{an optional value between 0 and 1 giving the maximal predicted marginal probability. If set to NULL (default), the algorithm will try to estimate it from the data.}

\item{start}{an optional list with elements named \code{beta}, \code{q}, and/or \code{mu1} giving starting values for estimation. If none or only some are specified, starting values are selected automatically.}

\item{control}{a list with parameters controlling the algorithm.}

\item{weight}{an optional vector specifying observation weights.}
}
\value{
an object of class \code{sprr} with the fitted model.
}
\description{
Fit semi-parametric relative risk model
}
