% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGLM.R
\name{spglm}
\alias{spglm}
\title{Fit semi-parametric GLM}
\usage{
spglm(
  formula,
  data,
  subset,
  weights,
  offset,
  link = "logit",
  mu0 = NULL,
  control = list(eps = 0.001, maxit = 100)
)
}
\arguments{
\item{formula}{a one-sided formula of the form \code{cbind(r, s) ~ predictors} where \code{r} and \code{s} give the number of responses and non-responses within each cluster, respectively (so the cluster size is \code{r+s}), and \code{predictors} describes the covariates.}

\item{data}{an optional matrix or data frame containing the variables in the formula \code{formula}. By default the variables are taken from \code{environment(formula).}}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{weights}{an optional vector specifying observation weights.}

\item{offset}{an optional vector specifying an offset term (predictor with a fixed coefficient of 1).}

\item{link}{a link function for the mean.}

\item{mu0}{an optional numeric value constraining the mean of the baseline distribution}

\item{control}{a list with parameters controlling the algorithm.}
}
\value{
an object of class \code{spglm} with the fitted model.
}
\description{
Fit semi-parametric GLM
}
